Changes by Mr. Shahid
Everything works fine and I made two changes. As we will overwrite it, I deleted existing tables(notebook is attached)
One thing we can do here. Define write location or base location in a variable which will be used to delete or write it 
inside tables. Do it as next step. 
The Second thing is that I extracted it from a live web address and also integrated it into a notebook. You can test it once and run
it to cross check if everything is ok. If you cannot do it because of bad internet connection or speed, just do it offline and test it.
In the end we need to delete the temp location where we are extracting files to process. Could you delete the temp location after
processing.